create post feature with cloudinary and create comments model

+++ controllers/api/PostController.js
@@ -1,10 +1,11 @@
 const {
   getAllPost,
-  getPost,
+  getPostById,
   createPost,
   updatePost,
   deletePost,
 } = require('../../models/PostModel.js')
+const { cloudinary } = require("../../utils/cloudinary.js");
 const apiResponse = require('../../utils/apiRespone.js')
 
 exports.getAllPost = async (req, res) => {
@@ -16,10 +17,10 @@ exports.getAllPost = async (req, res) => {
   }
 }
 
-exports.getPost = async (req, res) => {
+exports.getPostById = async (req, res) => {
   const id = parseInt(req.params.id)
   try {
-    const post = await getPost({ id })
+    const post = await getPostById({ id })
     res.status(200).json(apiResponse(true, 'get post data by id', post))
   } catch (error) {
     res.status(500).json(apiResponse(false, error.message, null, 500))
@@ -27,10 +28,12 @@ exports.getPost = async (req, res) => {
 }
 
 exports.createPost = async (req, res) => {
-  const { title, description, imageUrl } = req.body
+  const { title, description} = req.body
 
   try {
-    const posts = await createPost({ title, description, imageUrl, userId: req.session.userId })
+    const mediaUrl = req.file.path
+    const mediaName = req.file.filename
+    const posts = await createPost({ title, description, mediaUrl, publicId: mediaName, userId: req.session.userId })
     res.status(201).json(apiResponse(true, 'create post data', posts))
   } catch (error) {
     res.status(500).json(apiResponse(false, error.message, null, 500))
@@ -42,7 +45,16 @@ exports.updatePost = async (req, res) => {
   const { title, description, imageUrl } = req.body
 
   try {
-    const posts = await updatePost({ id, title, description, imageUrl })
+    const mediaUrl = req.file.path
+    const mediaName = req.file.filename
+    const existing = await getPostById({ id })
+    if (existing.user.id != req.session.userId ) return res.status(401).json(apiResponse(false, "permission required", null, 401))
+    if (!existing) return res.status(404).json(apiResponse(false, "Media not found", null, 404))
+
+    const posts = await updatePost({ id, title, description, mediaUrl, publicId: mediaName })
+    await cloudinary.uploader.destroy(existing.publicId, {
+      invalidate: true,
+    })
     res.status(201).json(apiResponse(true, 'update post data', posts))
   } catch (error) {
     res.status(500).json(apiResponse(false, error.message, null, 500))
@@ -50,9 +62,16 @@ exports.updatePost = async (req, res) => {
 }
 
 exports.deletePost = async (req, res) => {
+  const id = parseInt(req.params.id)
   try {
-    const id = parseInt(req.params.id)
+    const existing = await getPostById({ id })
+    if (existing.user.id != req.session.userId ) return res.status(401).json(apiResponse(false, "permission required", null, 401))
+    if (!existing) return res.status(404).json(apiResponse(false, "Media not found", null, 404))
+    
     const posts = await deletePost({ id })
+    await cloudinary.uploader.destroy(existing.publicId, {
+      invalidate: true,
+    })
     res.status(201).json(apiResponse(true, 'delete post data', posts))
   } catch (error) {
     res.status(500).json(apiResponse(false, error.message, null, 500))
